Move Pillar's merge_sources to salt.utils.dictupdate
